-Renamed soiling_model to heliosoil
