+++ analysis_scripts/raygen_analysis_heliostats.py
@@ -4,10 +4,10 @@ import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
 import scipy.stats as sps
-import soiling_model.fitting as smf
-import soiling_model.base_models as smb
-import soiling_model.utilities as smu
-from soiling_model.paper_specific_utilities import (
+import heliosoil.fitting as smf
+import heliosoil.base_models as smb
+import heliosoil.utilities as smu
+from heliosoil.paper_specific_utilities import (
     plot_for_paper,
     daily_soiling_rate,
     fit_quality_plots,
